Refactoring: каждая структура вынесена в отдельный файл со своими методами
